// src/api/generated/models/listAdminAppCasesRequestStatus.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ListAdminAppCasesRequestStatus = typeof ListAdminAppCasesRequestStatus[keyof typeof ListAdminAppCasesRequestStatus];


export const ListAdminAppCasesRequestStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN',
} as const;