// src/api/generated/models/listAuditRecordsRequestStatus.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ListAuditRecordsRequestStatus = typeof ListAuditRecordsRequestStatus[keyof typeof ListAuditRecordsRequestStatus];


export const ListAuditRecordsRequestStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN',
} as const;