// src/api/generated/models/listAppsRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ListAppsRequestSortOrder } from './listAppsRequestSortOrder';

export interface ListAppsRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  sortField?: string;
  sortOrder?: ListAppsRequestSortOrder;
  keyword?: string;
  status?: string;
  deployStatus?: string;
}