// src/api/generated/models/listAuditRecordsRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ListAuditRecordsRequestStatus } from './listAuditRecordsRequestStatus';

export interface ListAuditRecordsRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  status?: ListAuditRecordsRequestStatus;
}